Refactor SignIn screen to use dynamic fragments instead of activities: -SignIn screen was previously an Activity, now the Activity is basically just a container than connects two fragments, the SignIn fragment (all functionality moved from activity to fragment) and the Loading fragment (when a user attempts to sign in, a loading screen is shown to simulate a network delay before the sign in attempt is processed) 	-The activity implements two interfaces so that both fragments 	can communicatea back up to the activity -The Loading screen creates an AsyncTask that waits for two seconds and then returns the result to the activity -Upon config changes, the Loading AsyncTask gets cancelled and the result is returned earlier -The Loading fragment is not pushed to the back stack because the user should not be able to press the back button to go back to the SignIn fragment in the middle of loading/processing the SignIn attempt
Other Changes:
-In the SignIn fragment, when coming back from an activity, the text
remains the same as it was before leaving, except that the cursor is at
the beginning of the text. Place the cursor at the end of the text
instead, as expected by the user.
-Save email/password states the same as before except when the user
closes and reopens the app, the state should reset (which it didn't
previously). There is still a problem though, because it's only reset if
the user closes the app from the SignIn Activity, but not from the
Register activity (****FIX THIS IN FUTURE COMMIT).
-When going to the Register Activity from the SignIn Activity, instead
of creating regular intents both ways, use StartActivityForResult. This
is because otherwise, if you return from Register activity and then
change configuration, it will receive the Register intent each time you
change configuration, but if you call StartActivityForResult, it will
only return the result once