Save email/password state correctly
When leaving the SignIn fragment, the state of the email/password fields
(what was typed in there before leaving) should be saved, so that upon
return to the fragment, those fields are the same as when leaving.
-When leaving for the Loading fragment, save this state in
SharedPreferences, and upon returning from Loading fragment, reload and
this state back into the fields and clear from SharedPreferences
-When leaving for the Register Activity, send an intent with extras
indicating the state of the fields, and upon return from the Register
Activity, read the extras from the intent and reload the fields. Both of
these methods together ensure that when closing and reopening the app,
the state and fields are cleared, but when in the app, it will persist
	-Can't use just SharedPreferences because tough to discern
	between if returning from Register Activity or reopening the app