Add saving of all fields when registering and editing profile (Note: still need to log user out upon changing the password)
Email and password and display picture are already saved upon
registration, and displayed in Edit Profile Screen. Implement saving of
all the other fields (Name, Gender, Phone, Major, Class). Also make the
Save button work and actually save all change to the profile

-Also, fix bug associated with saving images (images should be saved in
the Edit Profile screen only if the image was changed, otherwise it will
overwrite its own image file (open input and output stream to it) and
result in a blank image because of data loss). Easy to check if image
has changed