Populate subtext in Manual Entry Activity screen with relevant info when inputting info in a dialog
When inputting info into a dialog, the subtext for that attribute should
contain the info you just inputted. The data is also stored as a string
in ManualEntryActivity

Fixed Bug: When inputting a period into Duration or Distance (the fields
that accept decimal numbers), it's accepted as a valid input ".", but
would cause an error when converting to double (for storing in the
database). Don't accept a period alone as an input

Fixed Bug: Only the field title was clickable (not the subtext),
resulting in a very small area of each field being clickable and
seemingly not responding half the time to touches. Place each field
(title and subtext and divider) into a LinearLayout and make the layout
clickable so that there is no area that doesn't respond to touch input