Fix: Database create statement doesn't have spaces between column name and data type
Update ExerciseEntry and its fields and add a couple of constructors.
Make the class be addable to the database, so change a couple of
variables to Strings to make it easier, etc.

In ManualEntryActivity, get the Privacy user setting to be able to pass
into the ExerciseEntry

When clicking the Save button, create an ExerciseEntry and add it to the
database

TLDR;Adding entries to the database now works, need to put database
operations on a separate thread and also need to bump the user back to
the Main screen after saving an entry