Add email field to database, so that every user has their own entries and it seems like they have their own database, but the History only queries the entries with the currently logged in user's email as a field -Change all database operations to take in the user's email -Also, capitalize all database fields
Change duration to a float (not sure why it should be int in the
database)

Fix Bug: When registering a user, there is an error ever since the
Register activity became multipurpose for both Register and Edit
Profile. For some reason, the part of the register activity that
registers a new user or saves a current user's new password in Edit
Profile only works for Edit Profile because when I made the change to
work for Edit Profile, I accidentally overwrote the part that worked for
Register, instead of having an else if that works for both. Update this
to work correctly